+++ lua/scnvim/udp.lua
@@ -3,36 +3,81 @@
 -- @author David Granström
 -- @license GPLv3
 
-local utils = require('scnvim/utils')
 local uv = vim.loop
-
-local M = {
-  udp = nil,
-  port = 0,
-}
+local help = require('scnvim/help')
+local utils = require('scnvim/utils')
+local M = {}
 
 local HOST = '127.0.0.1'
 local PORT = 0
 
+--- UDP handlers.
+-- Run the matching function in this table for the incoming 'action' parameter.
+-- @see on_receive
+local Handlers = {}
+
+--- Update status line widgets
+function Handlers.status_line(args)
+  if not args then return end
+  vim.api.nvim_set_var('scnvim_stl_widgets', args)
+end
+
+--- Print function signature
+function Handlers.method_args(args)
+  if not args then return end
+  print(args)
+end
+
+--- Open a help file
+function Handlers.help_open_file(args)
+  if not args then return end
+  help.open(args.uri, args.pattern)
+end
+
+--- Search for a method name
+function Handlers.help_find_method(args)
+  if not args then return end
+  help.handle_method(args.method_name, args.helpTargetDir)
+end
+
+--- Receive data from sclang
+function Handlers.eval(result)
+  assert(result)
+  if eval_callback then
+    eval_callback(result)
+    eval_callback = nil
+  end
+end
+
+--- Callback for UDP datagrams
+local function on_receive(err, chunk)
+  assert(not err, err)
+  if chunk then
+    local object = utils.json_decode(chunk)
+    assert(object, '[scnvim] Could not decode json chunk')
+    local func = Handlers[object.action]
+    assert(func, '[scnvim] Unrecognized handler')
+    func(object.args)
+  end
+end
+
 --- Start the UDP server.
--- @param on_receive UDP datagram handler.
-function M.start_server(on_receive)
+function M.start_server()
   local handle = uv.new_udp('inet')
   assert(handle, 'Could not create UDP handle')
   handle:bind(HOST, PORT, {reuseaddr=true})
-  handle:recv_start(on_receive)
+  handle:recv_start(vim.schedule_wrap(on_receive))
+  M.port = handle:getsockname().port
   M.udp = handle
-  -- let sclang know which port the server is running on
-  local port = handle:getsockname().port
-  M.port = port
-  utils.send_to_sc('SCNvim.port = '..port)
 end
 
 --- Stop the UDP server.
 function M.stop_server()
   if M.udp then
     M.udp:recv_stop()
-    M.udp:close()
+    if not M.udp:is_closing() then
+      M.udp:close()
+    end
     M.udp = nil
   end
 end
